refactor snapshot controller
