+++ pkg/virt-controller/watch/snapshot/util.go
@@ -40,6 +40,11 @@ var currentTime = func() *metav1.Time {
 	return &t
 }
 
+func timeTrack(start time.Time, name string) {
+	elapsed := time.Since(start)
+	log.Log.Infof("%s took %s", name, elapsed)
+}
+
 func cacheKeyFunc(namespace, name string) string {
 	return fmt.Sprintf("%s/%s", namespace, name)
 }
@@ -76,16 +81,6 @@ func newFreezingCondition(status corev1.ConditionStatus, reason string) snapshot
 		Reason:             reason,
 		LastTransitionTime: *currentTime(),
 	}
-}
-
-func findCondition(conditions []snapshotv1.Condition, c snapshotv1.Condition) bool {
-	for i := range conditions {
-		if conditions[i].Type == c.Type && conditions[i].Status == c.Status && conditions[i].Reason == c.Reason {
-			return true
-		}
-	}
-
-	return false
 }
 
 func updateCondition(conditions []snapshotv1.Condition, c snapshotv1.Condition, includeReason bool) []snapshotv1.Condition {
